Remove deltares logo and move sidebar logos to top
