+++ src/components/MetocSidebar.vue
@@ -1,16 +1,5 @@
 <template>
   <div style="height: 100%; display: flex; flex-direction: column">
-    <v-container>
-      <v-row>
-        <v-col>
-          <v-img class="ml-6" max-width="42" contain :src="minDefLogo" />
-        </v-col>
-        <v-spacer/>
-        <v-col>
-          <v-img class="ml-6" max-width="42" contain :src="navyLogo" />
-        </v-col>
-      </v-row>
-    </v-container>
     <TreeMenu
       ref="tree"
       :active.sync="activeNodes"
@@ -42,14 +31,6 @@ export default class MetocSidebar extends Mixins(PiRequestsMixin) {
 
   mounted() {
     this.fillNodes()
-  }
-
-  get navyLogo() {
-    return `${process.env.BASE_URL}images/navy_logo.png`
-  }
-
-  get minDefLogo() {
-    return `${process.env.BASE_URL}images/mindef_logo.png`
   }
 
   @Watch('categories')
